Generate HTML body

Generates HTML body.

Usage

html_body(countcheck_list, unit_df, title, column_headers, home_url = NULL)

Arguments

countcheck_list

Data to be shown in report. List of lists with two items named "caption" and "countcheck_df". The item countcheck_df in each of the nested lists is a data frame with columns y_new, ucl_partpool, and unit - as returned by select_for_report(). The item caption in each of the nested lists is used for the table caption.

unit_df

Data frame with columns unit, unit_group_name, unit_name, and unit_url (ignored if NA). Must contain exactly one record for each unit in countcheck_df dataframes in countcheck_list.

title

for HTML document

column_headers

Column headers for tables; vector with five headers - the elements of the vector must be named group, count, ucl, unit, and name

home_url

URL for link to "Home" included at top of HTML document (optional)

Value

String with HTML code for body