Generate HTML report

Generates HTML report for multiple tables as returned by select_for_report().

html_report(
  countcheck_list,
  unit_df,
  title = "Report",
  table_width_px = 1000,
  column_headers = c(group = "Group", count = "Count", ucl = "UCL", unit = "Unit", name =
    "Name"),
  charset = "utf-8",
  lang = "en",
  home_url = NULL,
  style = NULL
)

Arguments

countcheck_list

Data to be shown in report. List of lists with two items named "caption" and "countcheck_df". The item countcheck_df in each of the nested lists is a data frame with columns y_new, ucl_partpool, and unit - as returned by select_for_report(). Each countcheck_df must contain at least one row. The item caption in each of the nested lists is used for the table caption.

unit_df

Data frame with columns unit, unit_group_name, unit_name, and unit_url (ignored if NA). Must contain exactly one record for each distinct unit in countcheck_df dataframes in countcheck_list.

title

Title for HTML document

table_width_px

Table width in pixels (for all tables)

column_headers

Column headers for tables; vector with five headers - the elements of the vector must be named group, count, ucl, unit, and name

charset

Character set of data and HTML document (must contain ASCII as subset)

lang

Language of HTML document

home_url

URL for link to "Home" included at top of HTML document (optional)

style

Additional CSS rules (optional)

Value

String with HTML code for report