Generate HTML table

Generates HTML table.

Usage

html_table(
  countcheck_df,
  unit_df,
  caption,
  headers = c(group = "Group", count = "Count", ucl = "UCL", unit = "Unit", name =
    "Name")
)

Arguments

countcheck_df

Data frame with columns y_new, ucl_partpool, and unit - as returned by select_for_report

unit_df

Data frame with columns unit, unit_group_name, unit_name, and unit_url (ignored if NA). Must contain exactly one record for each unit in countcheck_df dataframes in countcheck_list.

caption

Caption for table

headers

Vector with five column headers - the elements of the vector must be named group, count, ucl, unit, and name

Value

String with HTML code for table