Selects rows with matching unit group

Selects rows with matching unit group - helper function used in select_for_report(). Condition: unit_group_name matches part of group name in unit_df for units in s.

matching_unit_group(s, unit_df, unit_group_name)

Arguments

s

Data frame of same type as returned by select_for_report()

unit_df

Data frame with columns unit, unit_group_name, unit_name, and unit_url (ignored if NA). Must contain exactly one record for each unit in s.

unit_group_name

Name of unit group for which data should be returned. Can be part of name (substring).

Value

Vector of type logical