No-pooling model

For the no-pooling model we estimate theta_i using only the observed counts of observational unit i. For each unit i our no-pooling estimate of theta_i is the ratio of the observed count values of interest y_i to the reference count values n_i.

Usage

theta_nopool(n, y)

Arguments

n

Previous reference count values (measure of exposure), must at least be 1

y

Previous count values of interest

Value

No-pooling estimates of theta